refactor : 더이상 사용하지 않는 주석처리된 코드 삭제, dto 위치 변경
